api_yamdb/requirements.txt
flake8==4.0.1
requests==2.26.0
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.9.4
python-dotenv==0.20.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.1
sqlparse==0.4.2